Fixed errors, added custom error handling, reorganised project to use flask blueprints and fixed issues with upload/download.
