+++ .gitignore
@@ -4,3 +4,4 @@ instance/
 
 __pycache__/
 env/
+musicapp/creds.py
