+++ musicapp/templates/collection.html
@@ -2,7 +2,7 @@
 {% block content %}
   {% if songs.items %}
     <legend class="border-bottom mb-4">
-      Your Songs
+      All Songs
     </legend>
     <article class="media content-section">
       <div class="media-body">
@@ -20,7 +20,7 @@
               <tr class="my-4">
                 <td scope="row">{{ ((songs.page - 1) * songs.per_page) + loop.index }}</td>
                 <td>
-                  <a href="{{ url_for('song', song_id=song.id) }}"
+                  <a href="{{ url_for('songs.song', song_id=song.id) }}"
                      class="song-body text-decoration-none">{{ song.title }}</a>
                 </td>
                 <td>{{ song.artist }}</td>
@@ -36,11 +36,13 @@
                 {% if page_num %}
                   {% if songs.page == page_num %}
                     <li class="page-item active">
-                      <a class="page-link" href="{{ url_for('collection', page=page_num) }}">{{ page_num }}</a>
+                      <a class="page-link"
+                         href="{{ url_for('main.collection', page=page_num) }}">{{ page_num }}</a>
                     </li>
                   {% else %}
                     <li class="page-item">
-                      <a class="page-link" href="{{ url_for('collection', page=page_num) }}">{{ page_num }}</a>
+                      <a class="page-link"
+                         href="{{ url_for('main.collection', page=page_num) }}">{{ page_num }}</a>
                     </li>
                   {% endif %}
                 {% else %}
