+++ musicapp/templates/errors/403.html
@@ -1,0 +1,7 @@
+{% extends "layout.html" %}
+{% block content %}
+  <div class="content-section">
+    <h1>You don't have permission access that page. (403) </h1>
+    <p>Please check your account and try again.</p>
+  </div>
+{% endblock content %}
