+++ musicapp/templates/errors/404.html
@@ -1,0 +1,7 @@
+{% extends "layout.html" %}
+{% block content %}
+  <div class="content-section">
+    <h1>Oops! Page Not Found(404)</h1>
+    <p>That page does not exist. Please try a different location.</p>
+  </div>
+{% endblock content %}
