+++ musicapp/templates/errors/500.html
@@ -1,0 +1,7 @@
+{% extends "layout.html" %}
+{% block content %}
+  <div class="content-section">
+    <h1>Something went wrong(500)</h1>
+    <p>We are facing some trouble on our end. Please try again after some time.</p>
+  </div>
+{% endblock content %}
