+++ musicapp/templates/home.html
@@ -20,7 +20,7 @@
               <tr class="my-4">
                 <td scope="row">{{ ((songs.page - 1) * songs.per_page) + loop.index }}</td>
                 <td>
-                  <a href="{{ url_for('song', song_id=song.id) }}"
+                  <a href="{{ url_for('songs.song', song_id=song.id) }}"
                      class="song-body text-decoration-none">{{ song.title }}</a>
                 </td>
                 <td>{{ song.artist }}</td>
@@ -36,11 +36,11 @@
                 {% if page_num %}
                   {% if songs.page == page_num %}
                     <li class="page-item active">
-                      <a class="page-link" href="{{ url_for('home', page=page_num) }}">{{ page_num }}</a>
+                      <a class="page-link" href="{{ url_for('main.home', page=page_num) }}">{{ page_num }}</a>
                     </li>
                   {% else %}
                     <li class="page-item">
-                      <a class="page-link" href="{{ url_for('home', page=page_num) }}">{{ page_num }}</a>
+                      <a class="page-link" href="{{ url_for('main.home', page=page_num) }}">{{ page_num }}</a>
                     </li>
                   {% endif %}
                 {% else %}
