+++ musicapp/templates/layout.html
@@ -13,16 +13,16 @@
           type="text/css"
           href="{{ url_for('static', filename='main.css') }}"/>
     {% if title %}
-      <title>Music App - {{ title }}</title>
+      <title>MusicLib - {{ title }}</title>
     {% else %}
-      <title>Music App</title>
+      <title>MusicLib</title>
     {% endif %}
   </head>
   <body>
     <header class="site-header">
       <nav class="navbar navbar-expand-md navbar-dark bg-steel fixed-top">
         <div class="container">
-          <a class="navbar-brand mr-4" href="{{ url_for('home') }}">Music Lib</a>
+          <a class="navbar-brand mr-4" href="{{ url_for('main.home') }}">MusicLib</a>
           <button class="navbar-toggler"
                   type="button"
                   data-toggle="collapse"
@@ -35,16 +35,16 @@
           <div class="collapse navbar-collapse" id="navbarToggle">
             <div class="navbar-nav mr-auto">
               {% if current_user.is_authenticated %}
-                <a class="nav-item nav-link" href="{{ url_for('home') }}">Home</a>
-                <a class="nav-item nav-link" href="{{ url_for('collection') }}">Collection</a>
-                <a class="nav-item nav-link" href="{{ url_for('upload') }}">Upload</a>
+                <a class="nav-item nav-link" href="{{ url_for('main.home') }}">Home</a>
+                <a class="nav-item nav-link" href="{{ url_for('main.collection') }}">Collection</a>
+                <a class="nav-item nav-link" href="{{ url_for('songs.upload') }}">Upload</a>
               {% endif %}
             </div>
             <!-- Navbar Right Side -->
             <div class="navbar-nav ms-auto">
               {% if current_user.is_authenticated %}
                 <form class="form-inline my-2 my-lg-0 d-flex justify-content-evenly"
-                      action="{{ url_for('search') }}"
+                      action="{{ url_for('songs.search') }}"
                       method="post">
                   {{ form.hidden_tag() }}
                   <input type="search"
@@ -55,11 +55,11 @@
                          required/>
                   <button type="submit" class="btn btn-dark">Search</button>
                 </form>
-                <a class="nav-item nav-link" href="{{ url_for('account') }}">Account</a>
-                <a class="nav-item nav-link" href="{{ url_for('logout') }}">Logout</a>
+                <a class="nav-item nav-link" href="{{ url_for('users.account') }}">Account</a>
+                <a class="nav-item nav-link" href="{{ url_for('users.logout') }}">Logout</a>
               {% else %}
-                <a class="nav-item nav-link" href="{{ url_for('login') }}">Login</a>
-                <a class="nav-item nav-link" href="{{ url_for('register') }}">Register</a>
+                <a class="nav-item nav-link" href="{{ url_for('users.login') }}">Login</a>
+                <a class="nav-item nav-link" href="{{ url_for('users.register') }}">Register</a>
               {% endif %}
             </div>
           </div>
