+++ musicapp/templates/register.html
@@ -58,7 +58,7 @@
   <div class="border-top pt-3">
     <small class="text-muted">
       Already Have An Account?
-      <a class="ml-2" href="{{ url_for('login') }}">Sign In</a>
+      <a class="ml-2" href="{{ url_for('users.login') }}">Sign In</a>
     </small>
   </div>
 {% endblock content %}
