+++ musicapp/templates/search_results.html
@@ -18,7 +18,7 @@
           <tr class="my-4">
             <td scope="row">{{ loop.index }}</td>
             <td>
-              <a href="{{ url_for('song', song_id=song.id) }}"
+              <a href="{{ url_for('songs.song', song_id=song.id) }}"
                  class="song-body text-decoration-none">{{ song.title }}</a>
             </td>
             <td>{{ song.artist }}</td>
