+++ musicapp/templates/song.html
@@ -11,7 +11,11 @@
             By <strong>{{ song.artist }}</strong> from <strong>{{ song.album }}</strong>
           </p>
         </div>
-        <audio controls autoplay controlslist="nodownload" class="w-100 mt-2 mb-2">
+        <audio controls
+               autoplay
+               loop
+               controlslist="nodownload"
+               class="w-100 mt-2 mb-2">
           <source src="{{ music }}" type="audio/mp3"/>
           <source src="{{ music }}" type="audio/mpeg"/>
           <source src="{{ music }}" type="audio/wav"/>
@@ -19,7 +23,7 @@
         </audio>
         <div class="btns">
           <a class="btn btn-outline-info"
-             href="{{ url_for('download', song_id = song.id) }}">Download</a>
+             href="{{ url_for('songs.download', song_id = song.id) }}">Download</a>
           {% if song.owner == current_user %}
             <button type="button"
                     class="btn btn-outline-danger m-2"
@@ -48,7 +52,7 @@
           </div>
           <div class="modal-footer">
             <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
-            <form action="{{ url_for('delete', song_id=song.id) }}" method='post'>
+            <form action="{{ url_for('songs.delete', song_id=song.id) }}" method='post'>
               <input type="submit" class="btn btn-danger" value="Delete" />
             </form>
           </div>
