+++ run.py
@@ -1,4 +1,6 @@
-from musicapp import app
+from musicapp import create_app
+
+app = create_app()
 
 if __name__ == '__main__':
     app.run(debug=True)
